[SOIN] Extrais le contenu de l'action d'attribution de rôle
...dans un composant dédié
